Implemented valuation for Property, Added mechanism to create blank database

+++ PortfolioManager.Repository/Entities/Account.cs
@@ -27,7 +27,7 @@ namespace Portfolio.BackEnd.Repository.Entities
         public int PortfolioValuationId { get; set; }
 
         public int PortfolioId { get; set; }
-        public int PropertyValue { get; set; }
+        public decimal PropertyValue { get; set; }
     }
